fixed repo structure

+++ cpp/include/libsumo_wrapper.h
@@ -81,7 +81,7 @@ public:
 
     inline void writeRow(const std::string& id, const double time) {
         const auto& pos = libsumo::Vehicle::getPosition(id);
-        writer_ << id << static_cast<uint64_t>(time) << libsumo::Vehicle::getSpeed(id) << libsumo::Vehicle::getAcceleration(id) << pos.x << pos.y << libsumo::Vehicle::getFuelConsumption(id) << libsumo::Vehicle::getLaneID(id) << parquet::EndRow;
+        writer_ << id << time << libsumo::Vehicle::getSpeed(id) << libsumo::Vehicle::getAcceleration(id) << pos.x << pos.y << libsumo::Vehicle::getFuelConsumption(id) << libsumo::Vehicle::getLaneID(id) << parquet::EndRow;
     }
 
     void setRowGroupSize(int64_t size) {
